chore: port assets generation to typescript
